API

Modules

AutoPkg.AutoPkg

Types and constants

Functions and macros

AutoPkg.@autopkg

Documentation

AutoPkg.AutoPkg — Module

AutoPkg

Automatic Julia package handling.

Mainly intended for standalone Julia notebooks.

AutoPkg.@autopkg — Macro

@autopkg expr

Looks for using and import statements in expr, and adds the required packages to a temporary Julia project environment before evaluating expr.

Automatically activates a temporary project environment on first use (no matter if expr contains any using or import statements or not), except when running in Gooogle Colab notebooks (which are isolated anyway).

@autopkg sets the environment variable $JULIA_PKG_PRESERVE_TIERED_INSTALLED to "true", to re-use packages already present in the Julia package depot if possible.

Example:

@autopkg begin
    using SomePackage, SomeOtherPackage
    import YetAnotherPackage

    if some_condition
        using OneMorePackage: some_feature
    end
end